Makefile added; targets for local action run and image building added; workflow for image building and publishing added; image building in action replaced with container pulling
