+++ Dockerfile
@@ -1,18 +1,14 @@
 FROM nikolaik/python-nodejs:python3.9-nodejs16
 
-ADD requirements.txt /requirements.txt
-ADD main.py /main.py
-ADD loc.py /loc.py
-ADD make_bar_graph.py /make_bar_graph.py
-ADD colors.json /colors.json
-ADD translation.json /translation.json
+ENV PYTHONUNBUFFERED 1
+ENV PYTHONDONTWRITEBYTECODE 1
 
-ENV PATH "$PATH:/home/root/.npm-global/bin"
+WORKDIR ./waka-readme-stats
 
-RUN python -m pip install --upgrade pip wheel setuptools
-RUN pip install -r requirements.txt
-RUN npm -g config set user root
-RUN npm i -g npm@next-8
-RUN npm i -g vega vega-lite vega-cli canvas
+ADD requirements.txt ./requirements.txt
+ADD Makefile ./Makefile
+ADD sources/* ./
 
-ENTRYPOINT ["python", "/main.py"]
+RUN make dependencies
+
+ENTRYPOINT python3 ./main.py
